requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
dj-database-url==2.3.0
dj-rest-auth==2.1.9
Django==5.1.3
django-allauth==0.50.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-filter==24.3
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
oauthlib==3.2.2
pillow==11.0.0
psycopg2==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
requests-oauthlib==2.0.0
sqlparse==0.5.2
whitenoise==6.8.2